requirements2.txt
# Needed for Fabric
ecdsa
pycrypto
fabric

# Needed for parsing fig.yml
PyYAML